unit2/requirements.txt
channels
appdirs
diff-match-patch
Django
django-appconf
django-cleanup
django-debug-toolbar
django-imagekit
django-import-export
django-resized
django-rest-auth
djangorestframework
olefile
packaging
pilkit
Pillow
pyparsing
pytz
six
sqlparse
tablib
asgi_redis
